// src/models/MarketPlaceContact.ts
/* WARNING: This file is auto-generated . Do not edit manually. */

export type MarketPlaceContact = {
  city: string | null;
  country: string | null;
  email: string;
  firstname: string;
  id: string;
  lastname: string;
  phone: string | null;
  province: string | null;
  street: string | null;
  title: string;
  zip: string | null;
};