// src/models/MarketPlacePartner.ts
/* WARNING: This file is auto-generated . Do not edit manually. */

export type MarketPlacePartner = {
  category: string | null;
  city: string | null;
  companyNationalIdentificationNumber: string | null;
  contact: string | null;
  country: string | null;
  description: string | null;
  language: string | null;
  legalForm: string | null;
  organisationDisplayName: string | null;
  organisationName: string | null;
  otherDetails: string | null;
  province: string | null;
  street: string | null;
  url: string | null;
  vat: string | null;
  zip: string | null;
};